migration/pyproject.toml
[project]
name = "migration"
version = "0.1.0"
description = "Database migration tool for noti-temply"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
packages = ["."]
